Rotations now have overlap checks
Added cloneShapeArray helper
Re-used collision with grid to test if a rotation will cause an overlap

Another problem. Rotating at the screen edge, can in rare cases cause
cells to be drawn outside of the grid array. Need to someonehow detect
this case, move the object until it is valid, and then do the legality
check